Decouple evaluator and pre_init files
